Move the SharedPropTypes outside of ui/shared since they're specific to the blueprint frontend models
