+++ app/ui/browser-blueprint/model/shared-prop-types.js
@@ -13,7 +13,7 @@ specific language governing permissions and limitations under the License.
 import { PropTypes } from 'react';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 
-import * as Model from '../../browser-blueprint/model';
+import * as Model from './';
 
 export const Page = PropTypes.instanceOf(Model.Page);
 export const Pages = ImmutablePropTypes.listOf(Page.isRequired);
