+++ app/ui/browser-blueprint/views/window/index.jsx
@@ -16,7 +16,7 @@ import { ipcRenderer } from '../../../../shared/electron';
 import * as gestures from '../../lib/gestures';
 
 import Style from '../../../shared/style';
-import * as SharedPropTypes from '../../../shared/model/shared-prop-types';
+import * as SharedPropTypes from '../../model/shared-prop-types';
 import BrowserChrome from './browser-chrome';
 import BrowserContent from './browser-content';
 import DeveloperBar from './developerbar';
